pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-governance-mcp"
version = "1.7.0"
description = "MCP server for AI governance document retrieval"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Jason"}
]
keywords = ["mcp", "ai", "governance", "claude"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "mcp==1.25.0",
    "pydantic==2.11.9",
    "pydantic-settings==2.11.0",
    "sentence-transformers==5.2.0",
    "rank-bm25==0.2.2",
    "numpy==1.26.4",
    "requests>=2.28.0",  # Required by sentence-transformers but not declared as of huggingface-hub>=1.0
]

[project.optional-dependencies]
context-engine = [
    "watchdog>=4.0.0",
    "tree-sitter>=0.21.0",
    "pymupdf>=1.24.0",
    "openpyxl>=3.1.0",
    "Pillow>=10.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "pip-audit>=2.7.0",
    "bandit>=1.7.0",
    "safety>=3.0.0",
    "ruff==0.12.0",
    "pre-commit>=4.0.0",
]

[project.scripts]
ai-governance-mcp = "ai_governance_mcp.server:main"
ai-governance-extract = "ai_governance_mcp.extractor:main"
ai-context-engine = "ai_governance_mcp.context_engine.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/ai_governance_mcp"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (require real ML models, deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "real_index: marks tests using production index",
]
filterwarnings = [
    "ignore::DeprecationWarning:sentence_transformers",
]
